# Make quotes sequences of graphemes, proper `fromLeft`, `count`