Also wrote a small HTML generator/DSL which uses all that. Maybe move it into tests later?

TODO for this PR:

* Test all available quote words. Just has to be done.
* Make some benchmarks vs. old quote implementation & publish them here.

I've already decided to merge no matter what the benchmarks say, but I still want to do them, simply out of curiosity.

There are some TODOs in the example I mentioned which I'll also need to have a look at.