@homonoidian 2022-08-28
* The ASCII-only sample string is the contents of `zigcodegen.cpp`, a 10 000 line long (and roughly 500 Kb in size), ASCII-only file (taken from Zig stage1 as you may have guessed).
* The clusters sample string is the contents of a file filled with a pattern of grapheme clusters followed by a few ASCII characters, and repeated a few thousand times. This makes the clusters sample string 260 Kb in size.

Here are the benchmark results I got. Note that it is impossible to benchmark grapheme clusters using OLD quote implementation, because it doesn't support them.

```
OLD: create                            675.96M (  1.48ns) (± 7.54%)    0.0B/op         1.01× slower
OLD: count                             678.18M (  1.47ns) (± 7.42%)    0.0B/op         1.00× slower
OLD: empty?                            677.80M (  1.48ns) (± 5.13%)    0.0B/op         1.00× slower
OLD: stitch to itself                  12.75k  ( 78.46µs) (± 5.73%)   969kB/op     53323.17× slower
OLD: stitch                            25.51k  ( 39.20µs) (± 7.33%)   484kB/op     26637.78× slower
OLD: at? 0                             64.85M  ( 15.42ns) (±17.01%)   16.0B/op        10.48× slower
OLD: at? 1000                          61.83M  ( 16.17ns) (±16.21%)   16.0B/op        11.11× slower
OLD: at? count - 1 (worst case)        61.60M  ( 16.23ns) (±16.19%)   16.0B/op        11.15× slower
OLD: slice_at? 0                      399.26M  (  2.50ns) (± 3.84%)    0.0B/op         1.72× slower
OLD: slice_at? count - 1 (worst case)  25.52k ( 39.18µs) (± 5.89%)   484kB/op      26919.09× slower
OLD: slice_at? count                  254.02M  (  3.94ns) (± 3.39%)    0.0B/op         2.70× slower

NEW ASCII ONLY: create                              38.82 ( 25.76ms) (± 0.89%)    0.0B/op  17696152.06× slower
NEW ASCII ONLY: count                             678.21M (  1.47ns) (± 7.26%)    0.0B/op         1.01× slower
NEW ASCII ONLY: empty?                            678.66M (  1.47ns) (± 7.32%)    0.0B/op         1.01× slower
NEW ASCII ONLY: stitch to itself                    8.08k (123.83µs) (± 5.01%)   969kB/op     85068.14× slower
NEW ASCII ONLY: stitch                             21.44k ( 46.64µs) (± 5.29%)   484kB/op     32043.29× slower
NEW ASCII ONLY: at? 0                             301.43M (  3.32ns) (± 3.16%)    0.0B/op         2.28× slower
NEW ASCII ONLY: at? 1000                          301.33M (  3.32ns) (± 3.21%)    0.0B/op         2.28× slower
NEW ASCII ONLY: at? count - 1 (worst case)        223.95M (  4.47ns) (± 2.83%)    0.0B/op         3.07× slower
NEW ASCII ONLY: slice_at? 0                       678.90M (  1.47ns) (± 4.57%)    0.0B/op         1.01× slower
NEW ASCII ONLY: slice_at? count - 1 (worst case)   20.19k ( 49.52µs) (± 5.43%)   484kB/op     34022.74× slower
NEW ASCII ONLY: slice_at? count                    78.58M ( 12.73ns) (± 2.28%)    0.0B/op         8.74× slower

NEW CLUSTERS: create                             175.17 (  5.71ms) (± 0.50%)    0.0B/op   3921810.24× slower
NEW CLUSTERS: count                             679.60M (  1.47ns) (± 4.72%)    0.0B/op         1.01× slower
NEW CLUSTERS: empty?                            679.24M (  1.47ns) (± 4.72%)    0.0B/op         1.01× slower
NEW CLUSTERS: stitch to itself                   20.61k ( 48.53µs) (± 6.21%)   511kB/op     33337.98× slower
NEW CLUSTERS: stitch                             37.35k ( 26.77µs) (± 6.33%)   255kB/op     18394.39× slower
NEW CLUSTERS: at? 0                               4.10M (244.05ns) (±10.03%)    112B/op       167.66× slower
NEW CLUSTERS: at? 1000                            7.48k (133.67µs) (± 2.43%)  10.5kB/op     91832.21× slower
NEW CLUSTERS: at? count - 1 (worst case)         157.49 (  6.35ms) (± 2.28%)   496kB/op   4362139.13× slower
NEW CLUSTERS: slice_at? 0                       687.00M (  1.46ns) (± 6.03%)    0.0B/op              fastest
NEW CLUSTERS: slice_at? count - 1 (worst case)   135.06 (  7.40ms) (± 3.54%)  1.22MB/op   5086589.70× slower
NEW CLUSTERS: slice_at? count                    79.14M ( 12.64ns) (± 1.51%)    0.0B/op         8.68× slower
```

* We can see that in case of an ASCII-only string, most time is spent on the creation of `Quote`. Creating a Quote from the outside is almost always an O(2N) (= O(N)) operation, unless you give it a few hints (like the amount of characters in the quote, and whether it is ASCII-only). **Implementing quote creation as one-pass (O(N)) instead of two-pass should increase the speed somewhat.** 
* For better performance, you should avoid creating a lot of quotes at runtime (e.g., by keeping one in a definition and/or using it in a read-only matter for as long as possible), because quotes cache the results of those two passes, and their descendant quotes respect the cache via some computation, meaning they won't do the two passes anymore. Thanks to the latter, quote creation should be faster than that presented in the benchmark results for Novika quote literals, because they are created only once (at slurp-time), and then their descendants are used.
* In general, the slowest parts of the old quote implementation and the new quote implementation when the string is ASCII-only are the same. In general, the slow down in these parts is about 30% (judging by how many times the new implementation is slower than the old one). In the new implementation, `slice_at? count - 1` has the same speed as the old implementation (at least in this particular case).
* Clusters create and stitch faster because the sample string size is smaller. If we were to benchmark two equivalent strings, we would see that the cluster implementation is slower. The general solution for optimizing quotes is to avoid grapheme computations  for as long as possible.
* `at 1000` case is slower in clusters because it's *always* O(N) where N = 1000 in this case, speaking in terms of loop iterations only (there is a lot of other stuff happening besides that). I believe that `byte_at` (which is used in the ASCII-only case) is O(1), being just a pointer + offset dereference (which I suppose it is).
* In clusters, the slowest cases are the worst cases `slice_at? count - 1` and `at? count - 1`. Again, this is expected, because both are O(N) + extra work.

The general conclusions are that a) we must avoid creating a lot of quotes, b) as language implementors, we should avoid grapheme computations for as long as possible, even at the cost of a single expensive computation (a single O(N) vs. a thousand of them makes a difference in the real world, even though it's still O(N)), and c) further optimization efforts are required.

Note that my comments, conclusions, and benchmark results must be taken with a grain of salt.